upscale支持float类型

+++ inference_gfpgan.py
@@ -24,7 +24,7 @@ def main():
     parser.add_argument(
         '-v', '--version', type=str, default='1.3', help='GFPGAN model version. Option: 1 | 1.2 | 1.3. Default: 1.3')
     parser.add_argument(
-        '-s', '--upscale', type=int, default=2, help='The final upsampling scale of the image. Default: 2')
+        '-s', '--upscale', type=float, default=2, help='The final upsampling scale of the image. Default: 2')
 
     parser.add_argument(
         '--bg_upsampler', type=str, default='realesrgan', help='background upsampler. Default: realesrgan')
